+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Cache calls to k8s and graphite from the API's content endpoint
 - Cache permissions in Auth
 - Increased default worker threads in Auth and API
+- Add mongodb index to fix disk space delete policy query performance
 
 ### Deprecated